remove tests from build, use JS migrations instead of ts, exclude needed helper functions from test utils
