+++ package.json
@@ -62,7 +62,9 @@
         "moment": "^2.29.4",
         "node-cron": "^3.0.2",
         "patch-package": "^6.5.1",
+        "pg": "^8.16.3",
         "rate-limit-redis": "^4.2.0",
+        "redis": "^4.7.1",
         "reflect-metadata": "^0.1.13",
         "siwe": "^1.1.6",
         "slugify": "^1.4.7",
@@ -229,8 +231,8 @@
         "db:migrate:run:local": "NODE_ENV=development npx typeorm-ts-node-commonjs migration:run -d ./src/ormconfig.ts",
         "db:migrate:revert:local": "NODE_ENV=development npx typeorm-ts-node-commonjs migration:revert -d ./src/ormconfig.ts",
         "db:migrate:create:local": "NODE_ENV=development npx typeorm-ts-node-commonjs  migration:generate migration/$MIGRATION_NAME -d ./src/ormconfig.ts",
-        "db:migrate:run:production": "NODE_ENV=production npx typeorm-ts-node-commonjs migration:run -d ./src/ormconfig.ts",
-        "db:migrate:rever:productiont": "NODE_ENV=production npx typeorm-ts-node-commonjs migration:revert -d ./src/ormconfig.ts",
+        "db:migrate:run:production": "NODE_ENV=production npx typeorm migration:run -d ./build/src/ormconfig.js",
+        "db:migrate:rever:productiont": "NODE_ENV=production npx typeorm migration:revert -d ./build/src/ormconfig.js",
         "prettify": "prettier --write '**/*.ts*' '**/*.test.ts*'",
         "db:migrate:seedToken:run": "NODE_ENV=development ts-node ./node_modules/typeorm/cli -f ./src/seedToken-ormconfig.ts migration:run",
         "db:migrate:seedToken:revert": "NODE_ENV=development ts-node ./node_modules/typeorm/cli -f ./src/seedToken-ormconfig.ts migration:revert",
